requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==4.0.0
django-countries==7.5.1
django-dotenv==1.4.2
django-htmx==1.14.0
djangorestframework==3.14.0
idna==3.4
pika==1.3.2
Pillow==9.5.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
